Added 3 new functions in preparation of doing price volatility calculations.
The first 3 of 4 functions do price analysis calculations. These calculations
will serve as inputs to the price volatility analysis calculation.

The first 3 functions are named as follows:
* `get_most_recent_equity_price`
* `get_equity_price_history`
* `calculate_price_differences`

Also, a 4th of 4 new function was added. This final function is a helper
function to better understand the underlying API response data structure.

The fourth function is named `see_data_structure`.